Fix NPE when myRepo hasn't been explicitly set, such as when starting from a keyboard search
