+++ src/main/java/gitflow/GitflowBranchUtilManager.java
@@ -4,6 +4,7 @@ import com.intellij.openapi.project.Project;
 import git4idea.GitUtil;
 import git4idea.repo.GitRepository;
 import gitflow.actions.GitflowActions;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.HashMap;
 import java.util.Iterator;
@@ -17,7 +18,12 @@ import java.util.List;
 public class GitflowBranchUtilManager {
     private static HashMap<GitRepository, GitflowBranchUtil> repoBranchUtilMap;
 
-    static public GitflowBranchUtil getBranchUtil(GitRepository repo){
+    @Nullable
+    static public GitflowBranchUtil getBranchUtil(@Nullable GitRepository repo){
+        if (repo == null) {
+            return null;
+        }
+
         return repoBranchUtilMap.get(repo);
     }
