+++ src/main/java/gitflow/actions/AbstractPublishAction.java
@@ -1,6 +1,8 @@
 package gitflow.actions;
 
 import git4idea.repo.GitRepository;
+import gitflow.GitflowBranchUtil;
+import org.jetbrains.annotations.NotNull;
 
 public abstract class AbstractPublishAction extends AbstractBranchAction {
     AbstractPublishAction(String actionName, BranchType type) {
@@ -12,11 +14,11 @@ public abstract class AbstractPublishAction extends AbstractBranchAction {
     }
 
     @Override
-    protected boolean isActionAllowedForBranch() {
-        if (!super.isActionAllowedForBranch()) {
+    protected boolean isActionAllowed(@NotNull GitflowBranchUtil branchUtil) {
+        if (!super.isActionAllowed(branchUtil)) {
             return false;
         }
-        
+
         return !branchUtil.isCurrentBranchPublished();
     }
 }
