+++ src/main/java/gitflow/actions/AbstractTrackAction.java
@@ -1,6 +1,8 @@
 package gitflow.actions;
 
 import git4idea.repo.GitRepository;
+import gitflow.GitflowBranchUtil;
+import org.jetbrains.annotations.NotNull;
 
 public abstract class AbstractTrackAction extends AbstractBranchAction {
     AbstractTrackAction(String actionName, BranchType type) {
@@ -12,7 +14,7 @@ public abstract class AbstractTrackAction extends AbstractBranchAction {
     }
 
     @Override
-    protected boolean isActionAllowedForBranch() {
+    protected boolean isActionAllowed(@NotNull GitflowBranchUtil branchUtil) {
         String prefix;
         switch (type) {
             case Feature:
@@ -28,7 +30,8 @@ public abstract class AbstractTrackAction extends AbstractBranchAction {
                 return false;
         }
 
-        boolean noRemoteBranches = branchUtil.getRemoteBranchesWithPrefix(prefix).isEmpty();
+        boolean noRemoteBranches =
+                branchUtil.getRemoteBranchesWithPrefix(prefix).isEmpty();
         boolean trackedAllBranches = branchUtil.areAllBranchesTracked(prefix);
 
         return noRemoteBranches == false && trackedAllBranches == false;
